new framework. got trades working
Migrated to use the ccxt framwork instead of python-binance. ccxt allows for expanding this project easily. Added docstrings for all of the utils methods. Updated the Trade table to be consistent with the parsed results from ccxt which are easier to reason about and deal with than the data directly from Binance. Should also be standardized. Also added a Fee table since ccxt reports that data in a separate dictionary